requirements.txt
feedparser==6.0.11
requests==2.31.0
Pillow==11.1.0
google-api-python-client==2.142.0
google-auth==2.34.0
python-dateutil==2.9.0
beautifulsoup4==4.12.3
lxml==5.3.0
numpy<2.0.0
opencv-python-headless>=4.9.0.80
geopy==2.4.1
dateparser==1.2.0
groq>=0.4.0
tenacity>=8.2.0
yfinance>=0.2.36
google-genai>=1.0.0
